pyproject.toml
[tool.poetry]
name = "wg_forge_backend"
version = "0.1.0"
description = ""
authors = ["ajib6ept <asalbert@narod.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.5"
psycopg2 = "^2.9.3"
prettytable = "^3.3.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}
flake8 = "^4.0.1"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
